requirement.txt
Package                       Version
----------------------------- -------
asgiref                       3.8.1
Django                        5.1
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
pip                           22.0.2
PyJWT                         2.9.0
setuptools                    59.6.0
sqlparse                      0.5.1
typing_extensions             4.12.2
